text and images cleanup
