--- components/01-atoms/images/icons/icons.stories.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-import icons from './icons.twig';
-
-import iconData from './icons.yml';
-
-/**
- * Storybook Definition.
- */
-export default { title: 'Atoms/Icons' };
-
-export const defaultIcons = () => <div dangerouslySetInnerHTML={{ __html: icons(iconData) }} />;
