--- components/01-atoms/text/headings/headings.stories.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-import headings from './headings.twig';
-
-/**
- * Storybook Definition.
- */
-export default { title: 'Atoms/Headings' };
-
-export const headingsExamples = () => <div dangerouslySetInnerHTML={{ __html: headings({}) }} />;
